Refactor : ModalProvider와 ModalRoot로 로직 분리
